Start showing the gnav much earlier   - Global navigation now shows as soon as the main body is built instead of waiting for the unav to load   - A function was created to predict the size of the unav based on the unav metadata and the existence of a cookie in order to prevent CLS   - a min-width was given to the brand to prevent CLS (in most cases)   - the load of attributes.js is now parallelized inside global-navigation.js instead of blocking the critical path inside utilities.js     - This required moving two functions that depended on attributes.js out of utilities.js   - Fixed unit tests
