+++ libs/blocks/global-navigation/global-navigation.css
@@ -153,6 +153,7 @@ header.global-navigation {
 .feds-brand-image.brand-image-only {
   height: 36px;
   width: auto;
+  min-width: 66px;
 }
 
 .feds-brand-image picture, .feds-brand-image img, .feds-brand-image svg,
