+++ libs/blocks/global-navigation/utilities/utilities.js
@@ -11,7 +11,6 @@ import {
   getFederatedUrl,
   getFedsPlaceholderConfig,
 } from '../../../utils/utils.js';
-import { processTrackingLabels } from '../../../martech/attributes.js';
 import { replaceKey, replaceText } from '../../../features/placeholders.js';
 import { PERSONALIZATION_TAGS } from '../../../features/personalization/personalization.js';
 
@@ -166,15 +165,6 @@ export const federatePictureSources = ({ section, forceFederate } = {}) => {
     });
 };
 
-export function getAnalyticsValue(str, index) {
-  if (typeof str !== 'string' || !str.length) return str;
-
-  let analyticsValue = processTrackingLabels(str, getConfig(), 30);
-  analyticsValue = typeof index === 'number' ? `${analyticsValue}-${index}` : analyticsValue;
-
-  return analyticsValue;
-}
-
 export function getExperienceName() {
   const experiencePath = getMetadata('gnav-source');
   const explicitExperience = experiencePath?.split('#')[0]?.split('/').pop();
@@ -244,19 +234,6 @@ export async function loadDecorateMenu() {
 
   resolve(menu.default);
   return cachedDecorateMenu;
-}
-
-export function decorateCta({ elem, type = 'primaryCta', index } = {}) {
-  const modifier = type === 'secondaryCta' ? 'secondary' : 'primary';
-
-  const clone = elem.cloneNode(true);
-  clone.className = `feds-cta feds-cta--${modifier}`;
-  clone.setAttribute('daa-ll', getAnalyticsValue(clone.textContent, index));
-
-  return toFragment`
-    <div class="feds-cta-wrapper">
-      ${clone}
-    </div>`;
 }
 
 let curtainElem;
@@ -659,6 +636,16 @@ export function getGnavHeight() {
   }
 
   return topHeight;
+}
+
+export function getUnavWidthCSS(unavComponents) {
+  const iconWidth = 32; // px
+  const flexGap = 0.25; // rem
+  const sectionDivider = getConfig()?.unav?.showSectionDivider;
+  const cartEnabled = /uc_carts=/.test(document.cookie);
+  const components = (cartEnabled ? unavComponents?.filter((x) => x !== 'cart') : unavComponents) ?? [];
+  const n = components.length ?? 3;
+  return `calc(${n * iconWidth}px + ${(n - 1) * flexGap}rem${sectionDivider ? ` + 2px + ${flexGap}rem` : ''})`;
 }
 
 /**
