+++ libs/utils/utils.js
@@ -1436,7 +1436,6 @@ async function loadPostLCP(config) {
     header.classList.add('gnav-hide');
     performance.mark('Gnav-Start');
     loadBlock(header);
-    header.classList.remove('gnav-hide');
   }
   loadTemplate();
   const { default: loadFonts } = await import('./fonts.js');
